+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT
 
 - Ensure explicit array join `on` fields are used by parent query
+- Prevent `changed` publish callback race condition by enforcing sequential processing
 
 ---